triangle index removed

+++ advmath/DefaultGeometryShader.hpp
@@ -10,7 +10,7 @@ public:
 
 
 public:
-    Triangle<Output> operator () (const Vertex& in0, const Vertex& in1, const Vertex& in2, unsigned int triangle_index) const
+    Triangle<Output> operator () (const Vertex& in0, const Vertex& in1, const Vertex& in2, size_t triangle_index) const
     {
         return Triangle<Output>{ in0, in1, in2 };
     }
